fix(2291): move fromBlock to TimeUtils and update CI test pipeline
