+++ packages/common/src/SolidityTestUtils.js
@@ -1,5 +1,3 @@
-const { UMA_FIRST_EMP_BLOCK } = require("./Constants.js");
-
 // Attempts to execute a promise and returns false if no error is thrown,
 // or an Array of the error messages
 async function didContractThrow(promise) {
@@ -147,15 +145,6 @@ async function mineTransactionsAtTime(web3, transactions, time, sender) {
   }
 }
 
-// Sets fromBlock to the value of an environment variable if one is set. This can be set to 0 to make tests work with Ganache, or any other value needed for a production script or bot.
-async function getFromBlock() {
-  if (process.env.FROM_BLOCK) {
-    return process.env.FROM_BLOCK;
-  } else {
-    return UMA_FIRST_EMP_BLOCK;
-  }
-}
-
 module.exports = {
   didContractThrow,
   mineTransactionsAtTime,
@@ -163,6 +152,5 @@ module.exports = {
   takeSnapshot,
   revertToSnapshot,
   stopMining,
-  startMining,
-  getFromBlock
+  startMining
 };
