+++ packages/common/src/TimeUtils.js
@@ -1,3 +1,5 @@
+const { UMA_FIRST_EMP_BLOCK } = require("./Constants.js");
+
 /**
  * @notice Return average block-time for a period.
  */
@@ -14,6 +16,16 @@ const averageBlockTimeSeconds = async (/* lookbackSeconds */) => {
   }
 };
 
+// Sets fromBlock to the value of an environment variable if one is set. This can be set to 0 to make tests work with Ganache, or any other value needed for a production script or bot.
+async function getFromBlock() {
+  if (process.env.FROM_BLOCK) {
+    return process.env.FROM_BLOCK;
+  } else {
+    return UMA_FIRST_EMP_BLOCK;
+  }
+}
+
 module.exports = {
-  averageBlockTimeSeconds
+  averageBlockTimeSeconds,
+  getFromBlock
 };
